Fixed autocommands not being in the vimrc group
The cause of this was that one of the files enabled ftdetect and syntax,
which meant that certain plugins were loaded after this file, and before
the rest of the rc files, which caused the augroup to be cleared.

The solution is to have a separate category: after, which is loaded at
the end of the vimrc, outside of the augroup.