Add path rewriting to serve-stuff
This enables you to serve a file from a different path than where it
actually is located, which can be useful to simulate the paths that
another build process would yield for the input files.

The way this works is you open the file at the "fake" path, and you add
a query parameter "rewritePath" that indicates the original path.

For example, say you have a file "foo/file.md", but you want to pretend
the name is "bar/file.md". The url would be
"bar/file.md?rewritePath=foo/file.md".