Rewrote the link script in python
This replaces the shell script with a python implementation, using only
the standard libraries.

This should be a bit more portable, because this has no dependency on realpath.

It should also be easier to maintain, due to just being writter better in general.

Additionally, the various config files have been merged into one, that
should be easier to read in general than the previous ones.