Extended the colors of the perms in ffind output
A directory that is readable, but not executable (listable) is now
yellow and underlined, instead of read. This makes it distinct from not
being readable at all, and communicates that it is readble, but with a
caveat.

A path that is both read and writable will (in addition to the exising
color, if any) be underlined if the sticky bit is set, and it is not
owned by the current user. As with the other use of underline, this
serves as a marker that there is a caveat to the indicated permissions.