Fixed the restart zsh keybind
It was broken due to a conflict with one of the new word keybinds. The
underlying problem was its dependence on actually sending the `exec zsh`
command as input to the shell, followed by a newline.

The new implementation instead directly manipulates the buffer and uses
zle calls, avoiding keybindings or existing input from interfering.