Simplified keybindings
Many of these used a loop to bind to both the lower and uppercase
version of the keybin. To start with, this doesn't even make sense for
bindings with control, as this is the same keybind. In addition, it just
needlessly uses two keybinds, of which I'm only using one anyway.