Rewrote ALEIgnore functionality
It has been improved in several significant ways.

You can now choose which errors to ignore, rather than always ignoring
all, through an FZF list of error codes + messages. Fixes #19.

It is now also possible to define that a comment has to appear on the
line above or below the offending line. This makes it possible to add
ignore-next-line or even ignore-area style comments. Fixes #18.

By using tcomment instead of commentary it also becomes possible to
specify the type of comment that should be used for each comment. This
is useful for some tools. For example, ESLint requires single-line
ignores to use single-line comments, and area-based ignores to use block
comments. Fixes #21.

It is now also possible to define multiple different ways in which a
problem can be ignored. This can be useful for tools that allow you to
either put the comment above or at the end of the offending line. You
even get a nice preview of what each option will look like. Fixes #20.