Removed the base64 plugin
I only needed a subset of it, and didn't use it otherwise. The mappings
were also not configurable.

Replaced it with a few simple heuristics that, in addition to
encoding/decoding, can also provide the raw shell commands used, which I
can use to make the fzf preview more cross-platform.