Fixed work transposition edge case
In the following examples | indicates the cursor position.

Starting with `echo 'a=|b'`, using
`forward-transpose-words-with-style-normal` resulted in `echo 'a='b|`.
`backward-transpose-words-with-style-normal` fails to undo this.

Starting with `echo 'a=b'|`, using
`backward-transpose-words-with-style-normal` resulted in `echo 'a=|'b`.
`forward-transpose-words-with-style-normal` fails to undo this.

This fixes both of these problems, causing the cursor to not move in the
given examples.