Improved performance of set-background
This is done by caching the output of feh, using this cached data in
subsequent runs, and updating the cache at the end of each run.

This means that, even though the overal run takes the same amount of
time, it gets to the point of actually changing the background quicker.

The downside of this is that changes to the backgrounds (particularly
additions or changes in existing backgrounds) will only be used in the
second run.