Fix linking to inside another target in link.py
For example, mapping a folder to `foo`, and another folder to `foo/bar`.
Previously this would result in a new directory `foo` being created, and
the mapping to `foo` going to `foo/foo` instead.

The way this is resolved is by only creating a directory if it is
missing at the time it is needed, by use of a small function that is
added to the top of the output script.