Sort the ln commands by target in link.py
This is to prevent issues with ordering. For example, take the following
ln commands:

ln bar ~/.foo/bar
ln foo ~/.foo

If performed using the above order, the first command will cause the
~/.foo directory to be created. This will cause the second command to
link to ~/.foo/foo instead, which was not intended.

After the sort, these commands will be in the reverse order, and this
problem will not appear.

(As a side note, this behaviour depends on the modified ln that creates
target directories as needed. Without this in place it would still not
work correctly, but it would error out instead of ending up in an
incorrect location.)