Simplified fzf preset code
It now uses an associative array, instead of a string.

The first advantage of this is that it doesn't require looping through
the entire list to lookup a preset. This makes the code simpler and
cleaner.

The second advantage is that preset names are now unique (as they were
always designed to be), and registering a preset overrides an existing
preset with this name. This allows reloading scripts defining presets,
which aids in developing presets.