Fixed downloads pipeline skipping the first item
This was due to the change that made it work in previews. Due to the
behaviour of zsh when the -c option is passed, the first item was
missed.

Relevant section from the manpage:

```
-c Take the first argument as a command to execute, rather than reading
commands from a script or standard input. If any further arguments are
given, the first one is assigned to $0, rather than being used as a
positional parameter.
```