Made a generic iconmenu using rofi
Inspired by https://gitlab.com/vahnrr/rofi-menus

I liked the menus this created, but did not like the amount of hardcoded
pixel sizes, and the fact that it did not handle different resolutions
particularly elegantly.

This requires installing rofi based on git, as it uses the `@media`
feature which has not been in any releases as of this commit.