Track pending changes in link.py
This allows it to properly handle some more complicated cases.

For example, a situation where the config for a folder is changed from
symlink to recurse would not be handled properly, because the files
inside the folder would be seen as identical. In this case it can now
track that the folder will be newly created, so the files will no longer
existing in that location.