Fixed multiple parents missing for link.py targets
If the parent of the target of a file doesn't exist, this will be queued
for creation. For example, if a file links to `~/.foo/bar`, and `~/.foo`
doesn't exist, a `mkdir ~/.foo` command will be added to the command
list.

This failed when the parent of the parent did not exist. For example, if
a file links to `~/.foo/bar/baz`, and `~/.foo` does not exist, this
resulted in an error.

This fixes this bug, so that this case would now result in both a `mkdir
~/.foo` and a `mkdir ~/.foo/bar` command.