Add run-with-sudo wrapper
Previously I used an alias to always run docker with sudo, but I too
frequently encounter projects that expect you to be able to run docker
without sudo. I will not be setting up my systems this was because of
the security problems with such an setup.

Because this often involves docker being ran from a script my alias was
not helping; having a `docker` and `docker-compose` wrapper that make
sure sudo is used does help as it will be used by these scripts as well.