Add wrapper scripts for neuron
Neuron has 3 options to install/run it:

1. Use nix to install it and its dependencies.
2. Use a provided bundle that includes all dependencies.
3. Use a prebuilt docker image.

1 requires setting up nix, a system-level dependency manager, which I'm
really not willing to do just to get a single executable.

2 works, but the wrapper runs in a chroot and only makes part of the
outside filesystem available. The part where my zettelkasten lives is
not included, unfortunately.

3 works perfectly but it has more overhead, and running a container
requires root. This is an issue for the neuron-vim plugin, which tries
to run neuron commands for some actions.

Because of this I ended up writing a system where a docker contains with
neuron runs with a script that communicates with some fifo files, which
are used by a wrapper script to trigger neuron commands and see the
results. This allows using 3 without having to run docker for each
command.