Add script to create a local install of nix
The idea of nix intruiged me, but the fact that it was a system-level
thing bothered me, as it meant I could not depend on it always being
available/installable.

I've now found a way around this using nix-user-chroot, which allows me
to have a nix install in my home directory, so I start playing around
with this.