Reorganize i3 workspaces
For one, the workspaces that are pretty much permanently on some of the
side screens have been moved towards the end. I don't need to switch
windows/focus to these as often as many of the other workspaces.

Two new workspaces were created on the `-` and `=` keys, because more =
better.

Finally, the workspaces are now shown as icons in polybar. This is a lot
more compact and it looks pretty. The layer workspaces don't have icons
assigned and are therefore not shown in the switcher.