Fix pinentry-auto on -echo TTYs
Previously pinenty-auto would always try to use a TTY pinentry when it
could determine the TTY it was running in. This caused issues when the
terminal in question was being being controlled by an application such
as neovim or tmux over ssh.

To avoid this problem pinentry-auto now checks to see if the TTY has the
echo characteristic disabled, and avoids using it if this is the case.
It also sanity-checks whether the TTY is even queryable at all before
this just in case, even though I can't think of a case where a defunct
TTY would be running pinentry-auto.