Keep old filters in mpv-lavfi-filter
Instead of replacing the filters the new lavfi filter is now added to
the end of the list of filters. The base assumption is still that the
new lavfi filter combines/replaces all preexisting filters, so all of
the other filters are disabled and have their name changed (so that
stuff like my cropControl script will ignore them), but they are kept.