Have separate types for different paths in link.py
There are now different types for paths corresponding to real files
(*RealPath) and for those corresponding to paths based on the config
(*VirtualPath). Similarly, there are different types for absolute paths
(Absolute*Path) and relative paths (Relative*Path).

This helps make it a bit more explicit what kind of path is used in what
location and how paths can be joined (if at all).