Be more careful in the pinentry tty check
Previously only the echo flag was checked, but I've encountered some
situations where the echo flag wasn't changed but other echo-like flags
were (like echoe and echok). I'm not sure if this intended or a bug, but
either way we have to deal with it, so consider those flags as well.