Optimzied mpv-playlist-edit
For each item in the edited list two jq calls were performed, one of
which would loop through the entire original playlist to find a match.
This is slow, which becomes very noticeable on large (> 1000 item)
playlists.

This has now been rewritten in a single jq call, which only loops over
the original playlist once to make a map, which is then used for all
items. This is _much_ faster, going from multiple minutes to < 10
seconds for a list with ~1300 items (all of which were kept).