Fix multiple edge cases in scandir in link.py
The first is that if the directory itself is from fs it would only look
at the fs and exclude any virtual additions.

The second was that deleted items were still returned in the scandir.
They did correctly have type NONE, but this was still a weird difference
between a file that had been deleted and one that never existed in the
first place, which is undesirable.