Update screenshot script
It now uses maim for most modes instead of scrot, which seems to work
slightly better. Most notably scrot's selection mode frequently resulted
in the selection rect being visible in the screenshot itself which is
undesirable.

In addition there are now (better) notifications for this process. If a
countdown is needed this is shown using notifications. There is a
notification once the screenshot has been made with two actions: default
(which opens the image) and delete.