Properly escape arguments to Rg, add RgRaw
Previously the arguments to Rg were shell-escaped, but they could still
be interpreted as an option if they started with `-`, so in some cases
manual 'escaping' was still needed.

The Rg command will now prefix the arguments with `--` to indicate
end-op-options, so the arguments will now always only be used as a
pattern.

A new command (RgRaw) does no escaping at all, for cases where passing
additional options to `rg` is desired.