Move SSH_AUTH_SOCK setup to zsh init
This because nix-darwin adjusts PATH in /etc/zshrc, not /etc/profile, so
any nix-managed programs won't be available in profile.d files.

This does mean that it will not be set properly for anything that uses
it that doesn't load my zshrc, but I can't come up with anything I use
that this applies to, so that should be fine. Probably. I hope.