Fix tmux-resurrect not restoring processes
Finally bothered to debug this, and it turns out that the mechanism used
to find the running commands (the `save commmand strategy`) that is used
by default will detect all processes running in the pane, even
background processes.

In cases where this results in multiple processes this will cause the
restore process to fail. I'm not sure whether it refuses to run at all
or just looks at the first process in these cases; the process that
should have been restored was never the first in the cases I've
observed. (The first one was usually histdb related.)

In addition this mechanism appeared to only look at direct children of
the pane, so cases where there was an extra layer (such as a nested
shell) did not function correctly either.

There might be cases where this is the desired behavious - there's no
telling what this intermediary process is/did and whether restoring will
work properly without it - but I observed several cases on my machine
where restoring these processed would be desirable, and none that would
be problematic.

To resolve this I've created my own strategy that looks at the process
tree, filters out everything that isn't a foreground process, and then
takes the last result (in the case of a command in a nested shell both
the process of the command and of the shell might be considered
foreground, and the process of the command is listed last in these
cases).