Use tinted-theming schemes for base16-colorscheme
The trigger for this is that one of the themes in the base16 listing has
disappeared, causing updates of the scheme listing to fail. The
tinted-theming scheme list cannot have that issue since it actually
includes the files for all schemes rather than just having links to
them. This also makes updates of the scheme listing _much_ faster.

I'm honestly not sure why I didn't go with this method in the first
place.