Restructure LSP configs using null-ls
This is a generic language server like EFM, but built into Neovim. This
makes it easier to integrate with other plugins, and should make it
easier to do some of the conditional stuff I'll need to support
differences in codestyle between my own Python code and the code I write
for work. It also has builtin support for many external tools, making it
easier to integrate them.