Only update set-background cache if feh succeeds
Recenlty the feh command that's used to build the cache has been having
issues, and this causes the cache to become invalid and subsequent
set-background calls to fail. This change makes it so that the cache
isn't overwritten if the new cache wasn't built successfully.

An alternative command to build the cache using imagemagick is:

``` zsh
bgpath="$HOME/.local/share/backgrounds"
for fn in "$bgpath"/*; do
	identify -format "$bgpath/%f\n%w\n%h\n" "$fn"
done >! "$bgpath/.cache"
```