Add base16 Kitty template
I'm not sure why this wasn't already in here. As part of the commit that
introduced the base16-colorscheme script it was already added to
`.gitignore` and `kitty.conf`, and it was already rendered on the
machine I developed this script on, so it clearly was at some point.

In most cases this isn't really necessary as the colors are also set by
ZSH, but a) Kitty might not run ZSH (either because it's started with
another command, or because the default shell is not set to ZSH), and b)
this only applies once ZSH has started, which causes a "flash" of
incorrect colors between the moment Kitty start rendering and the moment
ZSH applies the colors.