Remove snippet placeholders from <Tab>
This seems useful, but in practice I sometimes end up moving to the next
spot manually, and it'll remember the spots I haven't tabbed to at this
point. Then (many) minutes later I'll press tab to indent and end up
jumping to a (seemingly) random spot. This annoyance _far_ outweighs the
convenience of jumping to next spot (which I rarely actually did
anyway).