Remove windows-registry
I don't really expect to go back to Windows on any of my own machines at
this point, and I'm not going to (be allowed to) use registry tweaks on
a customer's machine, so there's no point in keeping these around.

Plus they're pretty old, so chances are some of them wouldn't even work
on Windows 11 anyway.