Deal with breakages from none-ls mass deprecation
This is just part 1 which simply disables the broken integrations,
replacements will be done in separate comments.

---

Issue for reference:
https://github.com/nvimtools/none-ls.nvim/issues/58.

What a mess. I get wanting to deprecate unused/unmaintained stuff, but
this doesn't seem like the best way to do it (as is also evident from
the community response to this issue).

For starters a deprecation period of 1-2 months is very short. I get
that the Neovim ecosystem is a pretty fast moving one, but it's also a
tool people depend on for work; breaking people's setup on such a short
notice sucks and does not inspire confidence in your plugin.

For some of the removed integrations the upgrade path is to a native LSP
that integrates the tool. This is fine-ish, and I totally get the
motivation. This LSP might also include more functionality than the tool
it replaces which needs to be configured/disabled so this switch can be
somewhat involved, but at least you'll end up with the same
functionality. (Except for the cases were the maintainers of none-ls
were wrong about the LSP offering all the features the removed tool had,
which is sloppy and gives the appearance they didn't do their due
diligence.)

In other cases though the "upgrade path" is to another tool entirely.
This is fine if the original tool has actually been
deprecated/abandoned, but this has also been done for some tools that
are stable/supported for which a "better" tool is suggested. This is
**INSANE**. People can't just decide to switch their work projects over
to another tool, and their companies are unlikely to be swayed by the
argument "someone online claims this is better and decided we should
switch". And even for projects where people can decide on the used tools
it's unreasonable and unrealistic to expect them to do so on your whim.

Seriously, one of the things that's been removed is **`flake8`**, which
is extremely mature, well maintained, and actively funded. It's had a
new major version released just months ago, and it's used by tens of
thousands of projects. The "upgrade path" for this is `ruff`. `ruff`
claims to be a drop-in replacement, but it's also still on a pre 1.0
version and doesn't have nearly the same market adoption. Many projects
will continue to use `flake8` for years to come, and there's absolutely
nothing wrong with that.

The biggest offenders in this category have ended up in plugins adding
extra none-ls sources so there are paths for people to get the
functionality they need back, but these removals just don't seem
appropriate. This all sends the signal that the none-ls project is only
interested in offering integrations with the tools the maintainers feel
people should be using rather than with the tools people are actually
using.

---

That being said, most of these changes ultimately work fine for my
purposes, and it's introduced me to some LSPs that I wasn't aware
existed, so I'll probably end up with a better setup than before.