New k9s base16 template
Based on the old theme, but with some colour swaps and tweaks. The main
motivation for this is that the theme I was using actually used the
wrong bases for all colours (e.g. red is 08, but it used 0D which is
actually blue). While adjusting this I ended up adding some more tweaks
to fit my preference & include elements that weren't themed yet.